Suppress warnings regarding the System.Graphics library:
warning CA1416: This call site is reachable on all platforms. 'Bitmap' is only supported on: 'windows'.

The reason for suppressing this warning is to allow the other errors and warnings to be noticed.

TODO (e.g. create ticket for work item): Choose an alternative for the System.Graphics (dependent on GDI+) library to support hosting the website on a non-windows OS. Note that .NET 6 (LTS) does not support this library.